Extended IntervalResolver
